requirements.txt
blinker==1.7.0
click==8.1.7
cx-Oracle==8.3.0
flask==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
werkzeug==3.0.2
zipp==3.18.1
